+++ Hibiscus.cabal
@@ -31,6 +31,6 @@ executable Hibiscus
                        -Wredundant-constraints
 Test-Suite test-bar
     type:           exitcode-stdio-1.0
-    hs-source-dirs: test
+    hs-source-dirs: test src
     main-is:        test-bar.hs
     build-depends:  base, containers, mtl
