--- test/Parser.hs
@@ -1,84 +1,0 @@
-{-# OPTIONS_GHC -w #-}
-{-# LANGUAGE CPP #-}
-{-# LANGUAGE MagicHash #-}
-{-# LANGUAGE BangPatterns #-}
-{-# LANGUAGE TypeSynonymInstances #-}
-{-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE PatternGuards #-}
-{-# LANGUAGE NoStrictData #-}
-#if __GLASGOW_HASKELL__ >= 710
-{-# LANGUAGE PartialTypeSignatures #-}
-#endif
-{-# LANGUAGE DeriveFoldable #-}
-module Parser where
-
--- import Data.ByteString.Lazy.Char8 (ByteString)
-import Data.Maybe (fromJust)
-import Data.Monoid (First (..))
-
-import Control.Applicative(Applicative(..))
-import Control.Monad (ap)
-
-type ByteString = String
-
--- * AST
-
-data Name a
-  = Name a ByteString
-  deriving (Foldable, Show)
-
-instance Eq (Name a) where
-  (Name _ x) == (Name _ y) = x == y
-
-data Type a
-  = TVar a (Name a)
-  | TPar a (Type a)
-  | TUnit a
-  | TList a (Type a)
-  | TArrow a (Type a) (Type a)
-  | TLit String
-  deriving (Foldable, Show)
-
-instance Eq (Type a) where
-  t1 == t2 = undefined
-
-data Argument a
-  = Argument a (Name a) (Maybe (Type a))
-  deriving (Foldable, Show)
-
-data Dec a
-  = Dec a (Name a) [Argument a] (Maybe (Type a)) (Exp a)
-  deriving (Foldable, Show)
-
-data Operator a
-  = Plus a
-  | Minus a
-  | Times a
-  | Divide a
-  | Eq a
-  | Neq a
-  | Lt a
-  | Le a
-  | Gt a
-  | Ge a
-  | And a
-  | Or a
-  deriving (Foldable, Show)
-
-data Exp a
-  = EInt a Integer
-  | EVar a (Name a)
-  | EString a ByteString
-  | EUnit a
-  | EList a [Exp a]
-  | EPar a (Exp a)
-  | EApp a (Exp a) (Exp a)
-  | EIfThen a (Exp a) (Exp a)
-  | EIfThenElse a (Exp a) (Exp a) (Exp a)
-  | ENeg a (Exp a)
-  | EBinOp a (Exp a) (Operator a) (Exp a)
-  | EOp a (Operator a)
-  | ELetIn a (Dec a) (Exp a)
-  deriving (Foldable, Show)
--- $Id: GenericTemplate.hs,v 1.26 2005/01/14 14:47:22 simonmar Exp $
-
